requirements.txt
snowflake-connector-python
requests
schedule
python-dotenv
pandas